Create a generic activity and an included layout
