+++ sample/src/main/java/com/github/pengrad/mapscaleview/sample/GoogleMapsActivity.kt
@@ -17,20 +17,9 @@ import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
 import kotlin.random.Random.Default.nextInt
 
-class GoogleMapsActivity : AppCompatActivity(), OnMapReadyCallback, OnCameraMoveListener, OnCameraIdleListener {
+class GoogleMapsActivity : GenericActivity(), OnMapReadyCallback, OnCameraMoveListener, OnCameraIdleListener {
 
     private lateinit var map: GoogleMap
-    private val scales by lazy {
-        listOf<MapScaleView>(
-            findViewById(R.id.scaleView),
-            findViewById(R.id.scaleViewMiles),
-            findViewById(R.id.scaleViewRtl),
-            findViewById(R.id.scaleViewBg),
-            findViewById(R.id.scaleViewBgFixed)
-        )
-    }
-    private val scaleView get() = scales[0]
-    private val density by lazy { resources.displayMetrics.density }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -52,69 +41,5 @@ class GoogleMapsActivity : AppCompatActivity(), OnMapReadyCallback, OnCameraMove
 
     override fun onCameraIdle() {
         update(map.cameraPosition)
-    }
-
-    private fun update(cameraPosition: CameraPosition) {
-        for (scale in scales) {
-            scale.update(cameraPosition.zoom, cameraPosition.target.latitude)
-        }
-    }
-
-    fun changeColor(view: View) {
-        val getColor = { Color.rgb(nextInt(255), nextInt(255), nextInt(255)) }
-        scales.forEach { it.setColor(getColor()) }
-    }
-
-    fun changeTextSize(view: View) {
-        scales.forEach { it.setTextSize((nextInt(10) + 12) * density) }
-    }
-
-    fun changeStrokeWidth(view: View) {
-        scales.forEach { it.setStrokeWidth((nextInt(3) + 0.5f) * density) }
-    }
-
-    fun changeMiles(view: View) {
-        scaleView.milesOnly()
-    }
-
-    private var isMeters = false
-    fun changeMeters(view: View) {
-        isMeters = !isMeters
-        if (isMeters) scales.forEach { it.metersOnly() }
-        else scales.forEach { it.metersAndMiles() }
-    }
-
-    fun changeSize(view: View) {
-        scaleView.layoutParams.width = nextInt(200) + 200
-        scaleView.requestLayout()
-    }
-
-    private var isDefaultFont = true
-    fun changeFont(view: View) {
-        isDefaultFont = if (isDefaultFont){
-            scales.forEach { it.setTextFont(Typeface.DEFAULT) }
-            false
-        } else {
-            scales.forEach { it.setTextFont(Typeface.DEFAULT_BOLD) }
-            true
-        }
-    }
-
-    private var outline = true
-    fun changeOutline(view: View) {
-        outline = !outline
-        scales.forEach { it.setOutlineEnabled(outline) }
-    }
-
-    private var direction = false
-    fun changeDirection(view: View) {
-        direction = !direction
-        scales.forEach { it.setExpandRtlEnabled(direction) }
-    }
-
-    var largeTiles: Boolean = false
-    fun changeTileSize(view: View) {
-        largeTiles = !largeTiles
-        scales.forEach { it.setTileSize(if (largeTiles) 512 else 256) }
     }
 }
