+++ sample/src/main/res/layout/activity_googlemaps.xml
@@ -13,124 +13,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        tools:ignore="HardcodedText">
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeColor"
-            android:text="color" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeTextSize"
-            android:text="text size" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeStrokeWidth"
-            android:text="stroke width" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeMiles"
-            android:text="miles" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeMeters"
-            android:text="meters" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeSize"
-            android:text="size" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeOutline"
-            android:text="outline" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeDirection"
-            android:text="direction" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeFont"
-            android:text="font" />
-
-        <Button
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:onClick="changeTileSize"
-            android:text="tile size"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top|end">
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewBgFixed"
-            android:layout_width="100dp"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            android:background="#3333" />
-
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewBg"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="8dp"
-            android:background="#3333"
-            app:scale_expandRtl="true" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="8dp"
-        android:orientation="vertical">
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewMiles"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:scale_miles="true"
-            app:scale_outline="true" />
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleViewRtl"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:scale_color="#dc5533"
-            app:scale_expandRtl="true" />
-
-        <com.github.pengrad.mapscaleview.MapScaleView
-            android:id="@+id/scaleView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:scale_color="#339cdc"
-            app:scale_textSize="20sp" />
-
-    </LinearLayout>
+    <include layout="@layout/sample_layout" />
 
 </FrameLayout>
